Start looking into 6502 emulation
Looked at some "easy" instructions (i.e. ones with only "implied"
addressing mode) and started to get a sense of what "decoding" an
opcode actually means.

To decode an opcode, we want to map that byte to a pair (instruction,
addressing_mode).

I've decided it'll be fun to generate a Rust `match` statement at
compile time, based on data from a table. This reference looks
unreasonably good, so probably we'll get the data from there somehow:
https://www.masswerk.at/6502/6502_instruction_set.html

Onwards and upwards!