Start implementing "mode->operand" step
I did some thinking just now and it'll probably help to introduce an
"Operand" type that basically represents a "location" (either in memory,
or the contents of the accumulator, or a literal constant).

The Operand can have methods to `get` and `set` its value. This'll make
it easier to write logic for the various instructions, in a way that's
"orthogonal" to the choice addressing mode.

The next step is to implement this idea.