Debug weird GETLN behaviour
It turns out the ROM code is storing the value $88 (which I think is
just an arbitrary value) to $c000 (the keyboard byte!)

iiuc, they're doing this b/c it let's them check what sort of hardware
they're on. But we've sloppily impl'd the keyboard memory-mapped IO as
part of RAM, so it's *actually getting set*, and messing up other
things.

For now, we just special-case it out, and then all the demos & games
I've written so far just work. Hurray! (The BASIC & sysmon repls still
crash though. But those are lower prio to debug.)

Next let's start thinking about what hardware features to impl, and it
what order. And let's make sure to have a plan for incrementally
impl'ing soft-switches that *doesn't* involve so much painful
debugging of ROM code... It was fun to do once, but I'd be happy if I
didn't have to do it again.